% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAdaptiveMatrix.R
\name{createAdaptiveMatrix}
\alias{createAdaptiveMatrix}
\title{Create adaptive matrix}
\usage{
createAdaptiveMatrix(M, n)
}
\arguments{
\item{M}{distance matrix}

\item{n}{bandwidth - number of nearest neighbours}
}
\value{
Distance matrix with finite distances only to n nearest neighbours
}
\description{
This function takes a full distance matrix and adaptive bandwidth and
returns a distance matrix where only as many nearest neighbours are given for each
point as the value of bandwidth. All other neighbours are given the distance Inf,
so that in the next step, as we calculate weighting, their weights will be 0.
}
\examples{
createAdaptiveMatrix(matrix(c(0,1,2,3,4, 1,0,5,4,1, 9,7,0,2,1, 5,4,6,0,4, 4,3,2,1,0), nrow=5, ncol=5, byrow=TRUE),2)
}
