% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMatrix.R
\name{DistMatrix}
\alias{DistMatrix}
\title{Distance matrix}
\usage{
DistMatrix(x, y)
}
\arguments{
\item{x}{x coordinate of points}

\item{y}{y coordinate of points}
}
\value{
A pairwise distance matrix
}
\description{
This function returns a pairwise distance matrix between all points in the dataset.
}
\examples{
DistMatrix(1:5,1:5)
}
