% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclDist.R
\name{EuclDist}
\alias{EuclDist}
\title{Euclidean distance function}
\usage{
EuclDist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x coordinate of point 1}

\item{y1}{y coordinate of point 1}

\item{x2}{x coordinate of point 2}

\item{y2}{y coordinate of point 2}
}
\value{
Euclidean distance between two points
}
\description{
This function returns Euclidean distance between points (x1,y1) and (x2,y2).
}
\examples{
EuclDist(1,1,2,2)
}
