% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_bandwidth.R
\name{gwr_bandwidth}
\alias{gwr_bandwidth}
\title{Optimal bandwidth}
\usage{
gwr_bandwidth(formula, data, kernel, weighting)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data frame of points with columns for regression formula, including x and y coordinates of all points}

\item{kernel}{kernel type (default is bisquare for now)}

\item{weighting}{weighting if "adaptive", calculate an adaptive kernel where the bandwidth (bw) corresponds to the number of nearest neighbours (i.e. adaptive distance); if "fixed", a fixed kernel is found (bandwidth is a fixed distance)}
}
\value{
A list with two elements: 1) optimal bandwidth, 2) a table with all bandwidths and AICcs
}
\description{
This function optimises bandwidth for a linear gwr model using AICc optimisation (i.e. optimal
bandwidth is the one where AICc is minimal).
}
