% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{gwr_basic}
\alias{gwr_basic}
\title{GWR basic function}
\usage{
gwr_basic(formula, data, bw, kernel, weighting)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data frame of points with columns for regression formula, including x and y coordinates of all points - these need to be projected coordinates!}

\item{bw}{bandwidth, fixed (geog dist) or adaptive (number of neighbours)}

\item{kernel}{kernel type: currently only 'bisquare'}

\item{weighting}{if "adaptive", calculate an adaptive kernel where the bandwidth (bw) corresponds to the number of nearest neighbours (i.e. adaptive distance); if "fixed", a fixed kernel is found (bandwidth is a fixed distance)}
}
\value{
Returns a list with two elements: the first element is a data frame with summary statistics, the second element is a data frame with original data and parameter estimates
}
\description{
This function calculates a linear gwr model
}
