% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_bisquare.R
\name{w_bisquare}
\alias{w_bisquare}
\title{Bi-square weight}
\usage{
w_bisquare(dd, bwd)
}
\arguments{
\item{dd}{distance from regression point in metres}

\item{bwd}{bandwidth distance in metres, either exact value (fixed kernel) or geog distance to furthest neighbour (adaptive kernel)}
}
\value{
A bi-square inverse distance weight
}
\description{
This function returns a bi-square inverse distance weight.
}
\examples{
w_bisquare(2,5)
w_bisquare(5,2)
w_bisquare(2,2)
}
